backend/pyproject.toml
[tool.poetry]
name = "gym"
version = "0.1.0"
description = ""
authors = ["adityanandanx <nandanaditya985@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
django-environ = "^0.11.2"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"
python-jose = "^3.3.0"
django-filter = "^24.3"
celery = "^5.4.0"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
pytest-cov = "^6.0.0"
markdown = "^3.7"
django-jazzmin = "^3.0.1"
psycopg2 = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
